refactor: organize code into packages
