+++ Dockerfile
@@ -12,7 +12,7 @@ COPY . .
 
 # Cross-compile for Linux amd64
 ENV CGO_ENABLED=0 GOOS=linux GOARCH=amd64
-RUN go build -o bot .
+RUN go build -o bot ./cmd/bot
 
 FROM alpine:latest
