+++ internal/tracker/commands.go
@@ -1,4 +1,4 @@
-package main
+package tracker
 
 import (
 	"fmt"
